Split grids into their own file (took me long enough)
